W-15854291: Refactored code to collapse hierarchy of FileSystem, AbstractFileSystem, AbstractExternalFileSystem, ExternalFileSystem and FtpFileSystem to  FileSystem and FtpFileSystem only.
